# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TRZWorld with Swagger",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "carlosbelchiorsilva@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/survivors/store": {
            "post": {
                "tags": [
                    "Survivors"
                ],
                "summary": "Create a new survivor",
                "operationId": "ecd9bb97de2fe457ef7feb7ce3273fca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "age",
                                    "gender_id",
                                    "latitude",
                                    "longitude",
                                    "inventory"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the survivor",
                                        "type": "string"
                                    },
                                    "age": {
                                        "description": "Age of the survivor",
                                        "type": "integer"
                                    },
                                    "gender_id": {
                                        "description": "Gender ID of the survivor",
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "description": "Latitude of the survivor's location",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "longitude": {
                                        "description": "Longitude of the survivor's location",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "inventory": {
                                        "description": "List of items in the survivor's inventory",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item": {
                                                    "description": "ID of the item in the survivor's inventory",
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the item in the survivor's inventory",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Details of the created survivor",
                                    "properties": {
                                        "id": {
                                            "description": "ID of the created survivor",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Name of the survivor",
                                            "type": "string"
                                        },
                                        "age": {
                                            "description": "Age of the survivor",
                                            "type": "integer"
                                        },
                                        "gender_id": {
                                            "description": "Gender ID of the survivor",
                                            "type": "integer"
                                        },
                                        "latitude": {
                                            "description": "Latitude of the survivor's location",
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "longitude": {
                                            "description": "Longitude of the survivor's location",
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "inventory": {
                                            "description": "List of items in the survivor's inventory",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "item_id": {
                                                        "description": "ID of the item in the survivor's inventory",
                                                        "type": "integer"
                                                    },
                                                    "quantity": {
                                                        "description": "Quantity of the item in the survivor's inventory",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid parameters provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating bad request",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/survivors/show/{survivor}": {
            "get": {
                "tags": [
                    "Survivors"
                ],
                "summary": "Get details of a specific survivor",
                "operationId": "55b41032466cd993ebca7fddfd7b7278",
                "parameters": [
                    {
                        "name": "survivor",
                        "in": "path",
                        "description": "ID of the survivor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Details of the survivor",
                                    "properties": {
                                        "id": {
                                            "description": "ID of the survivor",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Name of the survivor",
                                            "type": "string"
                                        },
                                        "age": {
                                            "description": "Age of the survivor",
                                            "type": "integer"
                                        },
                                        "gender_id": {
                                            "description": "Gender ID of the survivor",
                                            "type": "integer"
                                        },
                                        "latitude": {
                                            "description": "Latitude of the survivor's location",
                                            "type": "number",
                                            "format": "double"
                                        },
                                        "longitude": {
                                            "description": "Longitude of the survivor's location",
                                            "type": "number",
                                            "format": "double"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Survivor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating survivor not found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/survivors/update/{survivor}": {
            "post": {
                "tags": [
                    "Survivors"
                ],
                "summary": "Update the location of a survivor",
                "operationId": "48354327b9e43d5cdfd3867d5695d5d0",
                "parameters": [
                    {
                        "name": "survivor",
                        "in": "path",
                        "description": "ID of the survivor to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "latitude": {
                                        "description": "Latitude of the survivor's new location",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "longitude": {
                                        "description": "Longitude of the survivor's new location",
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful update",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid parameters provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating bad request",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}